fix fuzzy lookup on bilingual TMs
lookupTarget was taken to mean _only_ target in 42b750e, which is not the
intent.  This made testFuzzyLookupTargetOnlySearchesTargets pass, but not
for the right reason.  In fact, the test was failing because bilingual fuzzy
search was not taking keyLocale into account, so would find "A" in a
different locale.

This commit reverts 42b750e, takes keyLocale into account, and makes the
test more thorough regarding keyLocale and lookupTarget (and changes the
name).